refact test code
