+++ executor/executor_test.go
@@ -2,8 +2,14 @@ package executor
 
 import (
 	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"net/url"
+	"os"
 	"testing"
 
+	"github.com/screwdriver-cd/sd-cmd/config"
 	"github.com/screwdriver-cd/sd-cmd/screwdriver/api"
 )
 
@@ -20,6 +26,34 @@ const (
 	dummyFile        = "sd-step"
 	dummyDescription = "dummy description"
 )
+
+func setup() {
+	config.SDAPIURL = "http://fake.com/v4/"
+	config.SDStoreURL = "http://fake.store/v1/"
+	b, _ := ioutil.ReadFile("testdata/validShell.sh")
+	validShell = string(b)
+	b, _ = ioutil.ReadFile("testdata/invalidShell.sh")
+	invalidShell = string(b)
+}
+
+func makeFakeHTTPClient(t *testing.T, code int, body, endpoint string) *http.Client {
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(code)
+		w.Header().Set("Content-Type", "application/json")
+		fmt.Fprintf(w, body)
+	}))
+	tr := &http.Transport{
+		Proxy: func(req *http.Request) (*url.URL, error) {
+			if endpoint == "" {
+				return url.Parse(server.URL)
+			} else if req.URL.Path == endpoint {
+				return url.Parse(server.URL)
+			}
+			return req.URL, nil
+		},
+	}
+	return &http.Client{Transport: tr}
+}
 
 type dummySDAPIBinary struct{}
 
@@ -79,4 +113,10 @@ func TestNew(t *testing.T) {
 	if err == nil {
 		t.Errorf("err=nil, want error")
 	}
+}
+
+func TestMain(m *testing.M) {
+	setup()
+	ret := m.Run()
+	os.Exit(ret)
 }
