+++ screwdriver/api/api_test.go
@@ -21,24 +21,9 @@ const (
 	fakeJWT      = "fake-jwt"
 )
 
-func setEnv(key, value string) {
-	if cacheEnv == nil {
-		cacheEnv = make(map[string]string)
-	}
-	cacheEnv[key] = os.Getenv(key)
-	os.Setenv(key, value)
-}
-
 func setup() {
-	setEnv("SD_API_URL", fakeAPIURL)
-	setEnv("SD_API_TOKEN", fakeAPIToken)
-	config.LoadConfig()
-}
-
-func teardown() {
-	for key, val := range cacheEnv {
-		os.Setenv(key, val)
-	}
+	config.SDAPIURL = fakeAPIURL
+	config.SDAPIToken = fakeAPIToken
 }
 
 func makeFakeHTTPClient(t *testing.T, code int, body, endpoint string) *http.Client {
@@ -152,6 +137,5 @@ func TestCommand(t *testing.T) {
 func TestMain(m *testing.M) {
 	setup()
 	ret := m.Run()
-	teardown()
 	os.Exit(ret)
 }
